+++ morphocluster/frontend/src/mixins.js
@@ -1,3 +1,5 @@
+import globalState from "@/globalState.js";
+
 export default {
     data() {
         return {
@@ -34,5 +36,14 @@ export default {
             this.messages.unshift(msg);
             console.log(error);
         },
+        setBreadcrumbs(bc) {
+            globalState.setBreadcrumbs(bc);
+        },
+        setLoading(name) {
+            globalState.setLoading(name);
+        },
+        unsetLoading(name) {
+            globalState.unsetLoading(name);
+        }
     }
 }